Update
Implements Timelines to the EGM format using a similar spec to Objects.
I did not realize these weren't binary, I thought Objects were saved
just like Rooms, but no they are legible making them easy to fix if a
project corrupts, so I am deciding on this for the Timeline spec. Also
resolves some warnings and type safety, overall the warnings between LGM
and the plugin combined are down from 358 to 121, a large chunk of the
remaining ones are unenforced type safeties in ENIGMA Settings Frame.