Fix Include Population (#70)
Fixes population of the includes to the engine without making a binary compatibility breaking change. While I was in here, I also fixed the jar description because we renamed the repo. The repo is no longer under "Enigma" so a fresh clone is downloaded as "lgmplugin/". This makes it easier to build the plugin on a fresh clone.